pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "customagenthf"
version = "0.1.0"
description = "Reusable smolagents starter package with Zeerak feature agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "smolagents[toolkit]",
    "python-dotenv",
]

[project.optional-dependencies]
litellm = ["smolagents[litellm]"]
transformers = ["smolagents[transformers]"]

[project.scripts]
customagenthf = "customagenthf.cli:main"

[tool.setuptools.packages.find]
include = ["customagenthf", "customagenthf.*"]

[tool.setuptools.package-data]
"customagenthf.zeerak" = ["prompts/*.txt"]
